tc: move report building logic into `error.rs` from `reporting.rs`
